source/requirements.txt
Django==3.0.1
django-allauth==0.44.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
gunicorn==20.0.4
ipdb==0.13.8
psycopg2-binary==2.8.5
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
